Refactor: Use facade pattern (TitleMaker)
